fixes an indexing error in BuildLoop
BuildLoop tries to build closed line loops by walking along the graph of lines, trying to find the tightest-wrapping (in some direction, ccw or cw, I forget) lines it can find, hoping that they end up closing into a loop. If it actually manages to find a closed loop this way, there is one stray test at the end where it checks that the last line added and the first line added still form a convex polygon. Before this fix, "first line" accidentally referred to "previous line", which was true by fiat.

I am a little surprised I didn't have to change the order of the arguments to Cross to get it to work, but it seems to work in all previous good cases and fix the one bad one provided by effigy. So: OK!